........
Compiler warnings removed.

AddHeader.escript: rfind -> rFind
........


git-svn-id: https://svn.berlios.de/svnroot/repos/escript/branches/0.5.7.experimental@139 5693b1fc-3b75-4070-9b6b-3ce3692a40d5

+++ EScript/Runtime/Runtime.cpp
@@ -438,7 +438,7 @@ Object * Runtime::executeCurrentContext(bool markEntry) {
 			}
 			case Statement::TYPE_UNDEFINED:
 			default:{
-				std::cout << " #unimplementedStmt "<<stmt->getType();
+				std::cout << " #unimplementedStmt "<<static_cast<int>(stmt->getType());
 				resultRef=executeObj( stmt->getExpression() );
 				stmt = rtb->nextStatement();
 				break;
